+++ examples/llm-workflow-graph/.gitignore
@@ -1,0 +1,7 @@
+ 
+node_modules
+**/node_modules
+backend/node_modules
+backend/csv
+
+package-lock.json
