+++ examples/llm-workflow-graph/src/locales/en-US.ts
@@ -1,0 +1,3 @@
+export default {
+  Upload: 'Upload',
+};
