+++ examples/llm-workflow-graph/src/locales/index.tsx
@@ -1,0 +1,6 @@
+import enUS from './en-US';
+import zhCN from './zh-CN';
+export default {
+  'en-US': enUS,
+  'zh-CN': zhCN,
+};
