+++ examples/llm-workflow-graph/src/locales/zh-CN.ts
@@ -1,0 +1,3 @@
+export default {
+  Upload: '上传',
+};
