--- packages/studio-explore/server/readme.md
@@ -1,1 +1,0 @@
-## generate graph dataset
