--- packages/studio-website/src/pages/llm/index.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react';
-import ExploreApp from '@graphscope/studio-explore';
-interface ILLMAppProps {}
-
-const LLMApp: React.FunctionComponent<ILLMAppProps> = props => {
-  return <ExploreApp />;
-};
-
-export default LLMApp;
