+++ pnpm-workspace.yaml
@@ -1,5 +1,6 @@
 packages:
   - 'packages/*'
+  - 'examples/*'
 packageExtensions:
   '*':
     dependencies:
